requirements.txt
numpy>=1.24
scipy>=1.10
nibabel>=5.0
h5py>=3.8
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.2
pyyaml>=6.0
tqdm>=4.65

# Feature extraction (GPU recommended)
torch>=2.0
torchvision>=0.15
open-clip-torch>=2.20
transformers>=4.30

# Optional
nilearn>=0.10
jupyter>=1.0
pytest>=7.0
